SBERDOMA-62 move User.test.js to domains/user/schema/User.test.js
